AI_source_code/requirements.txt
absl-py==0.8.1
albumentations==0.5.2
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2021.5.30
cffi==1.14.0
charset-normalizer==2.0.4
click==8.0.1
configparser==5.0.2
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
docker-pycreds==0.4.0
efficientnet==1.0.0
entrypoints==0.3
gast==0.3.2
geopandas==0.9.0
gitdb==4.0.7
GitPython==3.1.18
google-pasta==0.1.8
grpcio==1.16.1
h5py==2.9.0
idna==3.2
image-classifiers==1.0.0
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.2
Jinja2==2.10.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.3.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
mock==3.0.5
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.5.1
notebook==6.0.2
numpy==1.19.1
olefile==0.46
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
pandas==1.1.5
pandocfilters==1.4.2
parso==0.5.2
pathtools==0.1.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==2.0.9
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.5.2
pyparsing==2.4.7
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.2
pyzmq==18.1.0
qtconsole==4.6.0
requests==2.26.0
scikit-image==0.17.2
scikit-learn==0.20.2
scipy==1.3.2
seaborn==0.11.2
segmentation-models==1.0.1
Send2Trash==1.5.0
sentry-sdk==1.3.1
Shapely==1.7.1
shortuuid==1.0.1
six==1.13.0
smmap==4.0.0
subprocess32==3.5.4
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tifffile==2020.9.3
torch==1.3.1
torchvision==0.4.2
tornado==6.0.3
tqdm==4.62.1
traitlets==4.3.3
typing-extensions==3.10.0.0
urllib3==1.26.6
wandb==0.12.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
